use nav model to store navigation state
